✨🔧 Add NumberOverflowException and NumberUnderflowException classes
NumberOverflowException and NumberUnderflowException classes have been created and added to the exception package. Then, made changes in the compileExpression function to throw these exceptions when the number exceeds limits. Also, modified the interpret function to handle long type variables.

+++ src/main/scala/Interpret.scala
@@ -31,7 +31,7 @@ import scala.util.{Success, Try}
   * println(result) // Outputs: Success(3)
   *   }}}
   */
-def interpret[A](expression: ast.Expression[A], environment: Map[String, Int]): Try[Int] = {
+def interpret[A](expression: ast.Expression[A], environment: Map[String, Long]): Try[Long] = {
   expression match {
     case ast.Num(value)   => Success(value)
     case ast.Var(sym)     => Try(environment(sym))
